Share layout tests
