+++ app/error.spec.ts
@@ -1,25 +1,14 @@
 import { expect } from "@playwright/test";
 import { test } from "../src/test/browser/fixtures";
 import { tenorFeaturedHandler } from "../src/test/api/tenor";
+import { testLayout } from "../src/test/browser/shared";
 
 test("initial render", async ({ page, baseUrl, requestInterceptor }) => {
   requestInterceptor.use(tenorFeaturedHandler(500));
 
   await page.goto(baseUrl);
 
-  await expect(page).toHaveTitle("gifs.run - the fastest way to share gifs");
-
-  const logo = page.getByRole("img", { name: "gifs.run logo" });
-  await expect(logo).toBeVisible();
-
-  const searchInput = page.getByRole("textbox", { name: "search" });
-  await expect(searchInput).toBeVisible();
-  await expect(searchInput).toBeEmpty();
-  await expect(searchInput).toBeFocused();
-  await expect(searchInput).toHaveAttribute("placeholder", "Search for a gif");
-
-  const searchButton = page.getByRole("button", { name: "Search" });
-  await expect(searchButton).toBeVisible();
+  await testLayout(page);
 
   const gifs = page.getByRole("button", {
     name: /Load full preview of gif with description/,
@@ -33,7 +22,4 @@ test("initial render", async ({ page, baseUrl, requestInterceptor }) => {
 
   const errorImage = page.getByRole("img", { name: "an error occured" });
   await expect(errorImage).toBeVisible();
-
-  const footer = page.getByRole("img", { name: "Powered by Tenor" });
-  await expect(footer).toBeVisible();
 });
